# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChargeGrid Intelligence API",
...
    "contact": {
      "name": "ChargeGrid Intelligence",
      "url": "https://www.goodwe.com/"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/platform/plans": {
      "get": {
        "tags": [
          "plataforma"
        ],
        "summary": "Planos",
        "description": "O que a GoodWe vende. Escopo de rede.",
        "operationId": "planos_api_v1_platform_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Planos Api V1 Platform Plans Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/platform/contract": {
      "get": {
        "tags": [
          "plataforma"
        ],
        "summary": "Contrato",
        "description": "Contrato desta praca, com as ultimas cobrancas.\n\nSeguranca: `site_id` vem do escopo, que para operador comum e' sempre o site\ndele - trocar o parametro na URL nao mostra o contrato do vizinho.",
        "operationId": "contrato_api_v1_platform_contract_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "admin: escolhe o site da rede",
              "title": "Site Id"
            },
            "description": "admin: escolhe o site da rede"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Contrato Api V1 Platform Contract Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "plataforma"
        ],
        "summary": "Contratar",
        "operationId": "contratar_api_v1_platform_contract_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "admin: escolhe o site da rede",
              "title": "Site Id"
            },
            "description": "admin: escolhe o site da rede"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Contratar Api V1 Platform Contract Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/contract/terminate": {
      "post": {
        "tags": [
          "plataforma"
        ],
        "summary": "Rescindir",
        "description": "Pede a rescisao. Dentro do prazo minimo, emite a multa junto.\n\nNao e' DELETE: nao apaga nem encerra no ato. O contrato corre ate o fim do\nciclo ja cobrado, e a resposta diz ate quando e quanto custa.",
        "operationId": "rescindir_api_v1_platform_contract_terminate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "admin: escolhe o site da rede",
              "title": "Site Id"
            },
            "description": "admin: escolhe o site da rede"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Rescindir Api V1 Platform Contract Terminate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/platform/invoices/{cobranca_id}/settle": {
      "post": {
        "tags": [
          "plataforma"
        ],
        "summary": "Dar Baixa",
        "description": "Baixa MANUAL, e so' de admin.\n\nNao ha integracao bancaria: a cobranca nasce aberta e alguem da rede confirma\no recebimento. Deixar o proprio estabelecimento declarar que pagou nao seria\nbaixa manual, seria baixa nenhuma.",
        "operationId": "dar_baixa_api_v1_platform_invoices__cobranca_id__settle_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cobranca_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cobranca Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Dar Baixa Api V1 Platform Invoices  Cobranca Id  Settle Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  },
...
}